+++ src/tests/solidityTests/assertEq/test.assertEq.multi.log.sol
@@ -1,0 +1,8 @@
+pragma solidity ^0.4.4;
+
+import "wafr/Test.sol";
+
+contract AssertEqMultiLoggingTest is Test {
+  function test_oneValidEqStringUint() {
+  }
+}