+++ src/tests/solidityTests/assertEq/test.assertEq.multi.sol
@@ -1,0 +1,8 @@
+pragma solidity ^0.4.4;
+
+import "wafr/Test.sol";
+
+contract AssertEqMultiTest is Test {
+  function test_oneValidEqBytesUint() {
+  }
+}