+++ src/tests/solidityTests/dirDepth/tests/test.secondDepth.sol
@@ -1,0 +1,9 @@
+pragma solidity ^0.4.4;
+
+import "wafr/Test.sol";
+
+contract DirDepthSecondTest is Test {
+  function test_oneValidAssertFalse() {
+    assertFalse(false, "should be false");
+  }
+}